Generate dependency tree in dot format
With very large projects, this initscript has some problems with regards
to memory usage since it is attempting to keep the entire dependency
tree in memory through any included configurations or attributes. In my
Gradle version at least it also appeared that the snykMergedDepsConf
would result in a cycle, but dependency cycles in Gradle are not
allowed within a configuration, so the combination of all configurations
should also not have any cycles.