fix: make init-script explicit for dep discovery
The `--init-script=path` argument was being set as a raw argument in the CLI, but recently due to the addition of the `--initScript` in gradle
for the reachability feature this has been made more explicit. In order to stay true to this convention and make it more evident to users
through documentation, we're making the `initScript` option explicit and being sent as a parameter for the `getAllDeps` function.