fix: transitives when using gradle-normalize-deps
Discovered a bug in the graph builder when using a coordinate map.

The id used to query the snykGraph needs to be the original id, not the id from the coordinate map.

We should have noticed in the test that chops off the transitive line abruptly. Now test/fixtures-with-wrappers/basic-with-deps/dep-graph-gradleNormalizeDeps-failed-search.json includes the full graph, even when resolution fails.